+++ ydb/services/metadata/secret/accessor/ya.make
@@ -1,0 +1,17 @@
+LIBRARY()
+
+SRCS(
+    secret_id.cpp
+)
+
+PEERDIR(
+    ydb/core/base
+    ydb/library/actors/core
+    ydb/library/aclib
+)
+
+END()
+
+RECURSE_FOR_TESTS(
+    ut
+)
