+++ ydb/services/metadata/secret/secret.cpp
@@ -32,28 +32,4 @@ IClassBehaviour::TPtr TSecret::GetBehaviour() {
     return TSecretBehaviour::GetInstance();
 }
 
-TString TSecretId::SerializeToString() const {
-    TStringBuilder sb;
-    sb << "USId:" << OwnerUserId << ":" << SecretId;
-    return sb;
-}
-
-TString TSecretIdOrValue::DebugString() const {
-    return std::visit(TOverloaded(
-        [](std::monostate) -> TString{
-            return "__NONE__";
-        },
-        [](const TSecretId& id) -> TString{
-            return id.SerializeToString();
-        },
-        [](const TSecretName& name) -> TString{
-            return name.SerializeToString();
-        },
-        [](const TString& value) -> TString{
-            return MD5::Calc(value);
-        }
-    ),
-    State);
-}
-
 }
