+++ ydb/services/metadata/secret/ya.make
@@ -21,6 +21,7 @@ PEERDIR(
     ydb/core/grpc_services/base
     ydb/core/grpc_services
     ydb/services/metadata/request
+    ydb/services/metadata/secret/accessor
 )
 
 END()
